SMeasure and closeness measures
