+++ src/CausalityTools.jl
@@ -21,9 +21,8 @@ module CausalityTools
     include("core.jl")
     include("methods/infomeasures/infomeasures.jl")
     include("methods/crossmappings/crossmappings.jl")
-    include("methods/JointDistanceDistribution.jl")
+    include("methods/closeness/closeness.jl")
     include("methods/correlation/correlation.jl")
-    #include("methods/smeasure/smeasure.jl")
     #include("example_systems/ExampleSystems.jl")
 
     include("utils/utils.jl")
