+++ src/deprecations/deprecations.jl
@@ -2,3 +2,4 @@ include("transferentropy.jl")  # remove this file when 2.0 is released.
 include("crossmap.jl") # remove this file when 2.0 is released.
 include("predictive_asymmetry.jl") # will be reintroduced in some 2.X release
 include("joint_distance_distribution.jl")
+include("smeasure.jl")
