+++ src/methods/closeness/closeness.jl
@@ -1,0 +1,2 @@
+include("JointDistanceDistribution.jl")
+include("SMeasure.jl")
