+++ test/independence/SurrogateTest/SurrogateTest.jl
@@ -1,0 +1,7 @@
+
+# Pairwise measures
+include("MutualInformation.jl")
+include("SMeasure.jl")
+
+# Conditional measures
+include("ConditionalMutualInformation.jl")
